+++ TheBindingOfAlba/Debug/TheBindingOfAlba.tlog/TheBindingOfAlba.lastbuildstate
@@ -1,0 +1,2 @@
+PlatformToolSet=v142:VCToolArchitecture=Native32Bit:VCToolsVersion=14.28.29333:TargetPlatformVersion=10.0.19041.0:
+Debug|Win32|C:\Users\Home\Source\Repos\TheBindingOfAlba\|
