+++ TheBindingOfAlba/Enemy.cpp
@@ -1,18 +1,7 @@
 #include "Enemy.h"
 
-Enemy::Enemy(float x, float y, Game* game)
-	: Actor("res/spider.png", x, y, 32, 32, game) {
-
-	vx = 1;
-
-	state = States::IDLE;
-
-	aDying = new Animation("res/spider_die.png", width, height,
-		128, 32, 6, 4, false, game);
-	aMoving = new Animation("res/spider_move.png", width, height,
-		128, 32, 6, 4, true, game);
-	animation = aMoving;
-
+Enemy::Enemy(string filename, float x, float y, int width, int height, Game* game) 
+	: Actor(filename, x, y, width, height, game) {
 }
 
 void Enemy::update() {
@@ -40,13 +29,13 @@ void Enemy::impacted() {
 	}
 }
 
-void Enemy::changeDirection(float x, float y) {
-	if (x < 0)
-		vx = 1;
-	else if (x > 0)
+void Enemy::changeDirection(int x, int y) {
+	if (this->x > x)
 		vx = -1;
-	if (y < 0)
-		vy = 1;
-	else if (y > 0)
+	else if (this->x < x)
+		vx = 1;
+	if (this->y > y)
 		vy = -1;
+	else if (this->y < y)
+		vy = 1;
 }
