+++ TheBindingOfAlba/Enemy.h
@@ -6,12 +6,11 @@
 class Enemy : public Actor
 {
 public:
-	Enemy(float x, float y, Game* game);
 	Enemy(string filename, float x, float y, int width, int height, Game* game);
 	void update();
 	void draw() override;
 	void impacted(); 
-	void changeDirection(float x, float y);
+	void changeDirection(int x, int y);
 
 	States state = States::IDLE;
 	Animation* aDying;
