+++ TheBindingOfAlba/EnemyChase.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "Enemy.h"
+
+class EnemyChase : public Enemy
+{
+public:
+	EnemyChase(float x, float y, Game* game);
+	void update();
+};
