--- TheBindingOfAlba/EnemyShooter.h
@@ -1,11 +1,0 @@
-#pragma once
-
-#include "Enemy.h"
-
-class EnemyShooter : public Enemy
-{
-public:
-	EnemyShooter(float x, float y, Game* game);
-
-};
-
